move interpolation tensor
